refactor: clean up old Activities tab and rename Activities Enhanced to Activities
- Remove old Activities tab from App.tsx navigation
- Rename Activities Enhanced container to Activities
- Delete redundant components: ActivityInfo, ActivityTimeline, ActivityEvents
- Remove unused SCSS files: activities.scss, activity-info.scss
- Update all imports and references to use new Activities container
- Clean up internationalization messages and remove old references
- Simplify tab structure to: Client Info, Activities, Events

This consolidates the Live Activities functionality into a single, enhanced tab
with better UX and removes duplicate/redundant code.
