+++ packages/ui-plugins/live-activities/src/App.tsx
@@ -14,7 +14,6 @@ import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import React, { useState } from 'react';
 import { PluginBridgeProvider } from '@assurance/plugin-bridge-provider';
 import Activities from './containers/activities';
-import ActivitiesEnhanced from './containers/activities-enhanced';
 import Events from './containers/events';
 import ClientInfo from './components/validation/client-info';
 import { defineMessages, IntlProvider, useIntl } from 'react-intl';
@@ -43,10 +42,6 @@ const messages = defineMessages({
     id: 'activities',
     defaultMessage: 'Activities'
   },
-  activitiesEnhanced: {
-    id: 'activitiesEnhanced',
-    defaultMessage: 'Activities Enhanced'
-  },
   events: {
     id: 'events',
     defaultMessage: 'Events'
@@ -73,18 +68,14 @@ function Inner() {
         <TabList UNSAFE_style={{ flex: 1 }}>
           <Item key="clientInfo">{formatMessage(messages.clientInfo)}</Item>
           <Item key="activities">{formatMessage(messages.activities)}</Item>
-          <Item key="activitiesEnhanced">{formatMessage(messages.activitiesEnhanced)}</Item>
           <Item key="events">{formatMessage(messages.events)}</Item>
         </TabList>
         <TabPanels  marginTop="size-200">
-          <Item key="activities">
-            <Activities />
-          </Item>
-          <Item key="activitiesEnhanced">
-            <ActivitiesEnhanced />
-          </Item>
           <Item key="clientInfo">
             <ClientInfo />
+          </Item>
+          <Item key="activities">
+            <Activities />
           </Item>
           <Item key="events">
             <Events />
