--- packages/ui-plugins/live-activities/src/containers/activities.scss
@@ -1,18 +1,0 @@
-.live-activity {
-  border: 1px solid var(--spectrum-global-color-gray-400);
-  padding: 1rem;
-  border-radius: 0.5rem;
-  background-color: white;
-
-  &:hover {
-    background-color: var(--spectrum-global-color-gray-200);
-  }
-}
-
-.status-light-compact {
-  align-items: center;
-
-  &::before {
-    margin: 0 !important;
-  }
-}
